Remove Partial<> for required attributes
First off, really like this factory tool!

I noticed whilst coding that VSCode was not warning me of errors when using the `create...Factory` functions and saw that the generated files wrap all the inputs like `Partial<...Input>`.
Any reason to go with wrapping the input type with a `Partial`? If there, would be happy to make it an option in the generator config instead of removing it completely.